Extract command palette fuzzy logic
